+++ source/databricks/calculation-engine/tests/integration/calculator_job/conftest.py
@@ -15,6 +15,7 @@
 from azure.identity import ClientSecretCredential
 from datetime import datetime
 from pyspark.sql import SparkSession, DataFrame
+from pyspark.sql.types import StructType
 import pyspark.sql.functions as F
 import pytest
 from unittest.mock import patch
@@ -26,6 +27,9 @@ from package.calculation.calculator_args import CalculatorArgs
 from package.codelists.process_type import ProcessType
 from package.constants import EnergyResultColumnNames, WholesaleResultColumnNames
 from package.infrastructure import paths
+from package.calculation_input.schemas import (
+    time_series_point_schema, metering_point_period_schema, charge_master_data_periods_schema, charge_price_points_schema, charge_link_periods_schema
+)
 
 
 @pytest.fixture(scope="session")
@@ -52,6 +56,73 @@ def calculator_args_wholesale_fixing(
     args.batch_id = C.executed_wholesale_batch_id
     args.batch_process_type = ProcessType.WHOLESALE_FIXING
     return args
+
+
+@pytest.fixture(scope="session")
+def grid_loss_responsible_test_data(
+    spark: SparkSession,
+    test_files_folder_path: str,
+) -> DataFrame:
+    return spark.read.csv(f"{test_files_folder_path}/GridLossResponsible.csv", header=True)
+
+
+@pytest.fixture(scope="session")
+def energy_input_data_written_to_delta(
+    spark: SparkSession,
+    test_files_folder_path: str,
+    calculation_input_path: str
+) -> None:
+    # Metering point periods
+    _write_input_test_data_to_table(
+        spark,
+        file_name=f"{test_files_folder_path}/MeteringPointsPeriods.csv",
+        table_name=paths.METERING_POINT_PERIODS_TABLE_NAME,
+        schema=metering_point_period_schema,
+        table_location=f"{calculation_input_path}/metering_point_periods"
+    )
+
+    # Time series points
+    _write_input_test_data_to_table(
+        spark,
+        file_name=f"{test_files_folder_path}/TimeSeriesPoints.csv",
+        table_name=paths.TIME_SERIES_POINTS_TABLE_NAME,
+        schema=time_series_point_schema,
+        table_location=f"{calculation_input_path}/time_series_points"
+    )
+
+
+@pytest.fixture(scope="session")
+def price_input_data_written_to_delta(
+    spark: SparkSession,
+    test_files_folder_path: str,
+    calculation_input_path: str
+) -> None:
+    # Charge master data periods
+    _write_input_test_data_to_table(
+        spark,
+        file_name=f"{test_files_folder_path}/ChargeMasterDataPeriods.csv",
+        table_name=paths.CHARGE_MASTER_DATA_PERIODS_TABLE_NAME,
+        schema=charge_master_data_periods_schema,
+        table_location=f"{calculation_input_path}/charge_master_data_periods"
+    )
+
+    # Charge link periods
+    _write_input_test_data_to_table(
+        spark,
+        file_name=f"{test_files_folder_path}/ChargeLinkPeriods.csv",
+        table_name=paths.CHARGE_LINK_PERIODS_TABLE_NAME,
+        schema=charge_link_periods_schema,
+        table_location=f"{calculation_input_path}/charge_link_periods"
+    )
+
+    # Charge price points
+    _write_input_test_data_to_table(
+        spark,
+        file_name=f"{test_files_folder_path}/ChargePricePoints.csv",
+        table_name=paths.CHARGE_PRICE_POINTS_TABLE_NAME,
+        schema=charge_price_points_schema,
+        table_location=f"{calculation_input_path}/charge_price_points"
+    )
 
 
 @pytest.fixture(scope="session")
@@ -116,3 +187,24 @@ def wholesale_fixing_wholesale_results_df(
 ) -> DataFrame:
     results_df = spark.read.table(f"{paths.OUTPUT_DATABASE_NAME}.{paths.WHOLESALE_RESULT_TABLE_NAME}")
     return results_df.where(F.col(WholesaleResultColumnNames.calculation_id) == C.executed_wholesale_batch_id)
+
+
+def _write_input_test_data_to_table(
+        spark: SparkSession,
+        file_name: str,
+        table_name: str,
+        table_location: str,
+        schema: StructType,
+) -> None:
+    spark.sql(f"CREATE DATABASE IF NOT EXISTS {paths.INPUT_DATABASE_NAME}")
+    spark.sql(f"CREATE TABLE IF NOT EXISTS {paths.INPUT_DATABASE_NAME}.{table_name} USING DELTA LOCATION '{table_location}'")
+
+    df = spark.read.csv(
+        file_name,
+        header=True,
+        schema=schema,
+    )
+
+    df.write.format("delta").mode("overwrite").option("overwriteSchema", "true").saveAsTable(
+        f"{paths.INPUT_DATABASE_NAME}.{table_name}"
+    )
